first commit simple chat
