+++ .gitignore
@@ -1,1 +1,2 @@
 node_modules/
+/simple-chat/node_modules/
