+++ homework1/correctSentence.test.js
@@ -1,0 +1,7 @@
+import correctSentence from './correctSentence';
+
+test('returns correct sentence', () => {
+  expect(correctSentence("greetings, friends")).toBe("Greetings, friends.")
+  expect(correctSentence("Greetings, friends")).toBe("Greetings, friends.")
+  expect(correctSentence("Greetings, friends.")).toBe("Greetings, friends.")
+})
