+++ simple-chat/src/index.css
@@ -1,0 +1,197 @@
+html {
+    box-sizing: border-box;
+}
+
+*,
+*::before,
+*::after {
+    box-sizing: inherit;
+}
+
+
+body {
+    margin: 0;
+    background-color: #fff;
+    color: #333;
+    font-size: 15px;
+    font-family: Helvetica, Arial, sans-serif;
+    font-weight: 400;
+    line-height: 20px;
+    display: flex;
+    justify-content: row;  
+}
+
+button{
+    border: none;
+    background-color: transparent;
+    cursor: pointer;
+}
+
+.chat{
+    height: 100vh;
+    width: 100vw;
+    background-color: #2e2e39;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+}
+.sidebar{
+    height: 100vh;
+    width: 80px;
+    background-color: #120d14; 
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-around;
+}
+.icon {
+    font-size: 48px; 
+    color: #ffffffad;
+}
+.main_header{
+    display: flex;
+    align-items: center;
+    height: 70px;
+    background-color: #686a77;
+    box-shadow: 0px 5px 10px  3px #120d14;
+    font-size: 20px;
+    line-height: 25px;
+    z-index: 1;
+}
+img{
+    width: 60px;
+}
+.img-icon{
+    width: 60px; height: 60px;
+    border-radius: 50%;
+    overflow: hidden;
+    margin-left: 10px;
+    margin-right: 10px;
+}
+
+.person-name{
+    font-weight: bold; 
+    color: #fff;
+    font-size: 28px;
+
+}
+
+.form-input{
+    height: 60px;
+    font-size: 20px;
+    border-radius: 40px;
+    background-color: #fff;
+    color: #120d14;
+    padding-left: 15px;
+}
+.form {
+    
+    width: 100%;
+    display: inline-block;
+    
+}
+
+input[type=submit] {
+    visibility: collapse;
+}
+
+input {
+    border: 0;
+    outline: none;
+    width: calc(100% - 2px);
+    
+}
+.message {
+    display: inline-block;
+    margin-bottom: 8px;
+    padding: 16px;
+    border-radius: 15px;
+    background-color: #b7c2d4;
+    box-shadow: 2px 5px 10px 3px #120d14;;
+    color: #120d14;
+    font-size: 20px;
+    line-height: 24px;
+    text-align: start;
+    hyphens: auto;
+}
+.message-p{
+    background-color: #c3b1bd;
+}
+.time{
+    color: #ffffff9f;
+    font-size: 13px;
+    text-align: end;
+}
+.time-p{
+    color: #ffffff9f;
+    font-size: 13px;
+    text-align: start;
+}
+.name{
+    color: #ffffff;
+    font-size: 20px;
+    text-align: end;
+    margin-bottom: 5px;
+    font-weight: bold; 
+}
+.name-p{
+    color: #ffffff;
+    font-size: 20px;
+    text-align: start;
+    margin-bottom: 5px;
+    font-weight: bold; 
+
+}
+.container-message{
+    /* align-self: flex-end; */
+    margin-bottom: 10px;
+    max-width: 300px;
+    justify-content: end;
+}
+
+
+
+.container-chat{
+    height: 92%;
+    display: flex;
+    flex-direction: column;
+    overflow: auto;
+}
+.container-person{
+    padding: 13px 20px 10px 0;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+
+}
+.container-message-person{
+    display: flex;
+    flex-direction: row;
+}
+.container-me{
+    padding: 5px 20px 10px 20px;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+
+}
+.lower{
+    display: flex;
+    flex-direction: row;
+    box-shadow: 0px -5px 10px  3px #120d14;
+    align-items: center;
+    padding: 5px;
+    
+}
+.container-chat::-webkit-scrollbar {
+    width: 12px;               
+}
+.container-chat::-webkit-scrollbar-thumb {
+    background-color: #ffffff67;   
+    border-radius: 20px;       
+    border: none  
+}
+
+
