+++ simple-chat/src/index.html
@@ -1,0 +1,132 @@
+<!DOCTYPE html>
+<html>
+    <head lang="en">
+        <meta charset="UTF-8">
+        <title>Simple chat</title>
+        <link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined" rel="stylesheet" />
+
+    </head>
+    <body>
+        <style>
+
+        </style>
+        <div class="sidebar">
+            <button><span class="material-symbols-outlined icon">menu</span></button>
+            <button><span class="material-symbols-outlined icon">chat_bubble</span></button>
+            <button><span class="material-symbols-outlined icon">group</span></button>
+            <button><span class="material-symbols-outlined icon">settings</span></button>
+
+        </div>
+        
+        
+        
+        <div class="chat">
+            
+            <div class="main_header">
+                <div class="img-icon"><img  src="../img/girl.jpg" alt=""></div>
+                <p><span class="person-name">Дарья Петрова</span> <br>Онлайн </p>
+                
+            </div>
+            
+            
+
+            <div class="container-chat">
+                <div class="container-person">
+                    <div class="container-message-person">
+                        <div class="img-icon"><img  src="../img/girl.jpg" alt=""></div>
+                        <div class="message-person">
+                            <div class="name-p">Дарья Петрова</div>
+                            <div class="message message-p">Привет, как твои дела?</div>
+                            <div class="time-p">15:20 25.09.2024</div>
+                        </div>        
+                    </div>
+
+                </div>
+                <div class="container-me"></div>
+
+            </div>
+                
+            
+            <div class="lower">
+                <button><span class="material-symbols-outlined icon">sentiment_satisfied</span></button>
+                
+                <form class="form" action="/">
+                    <input class="form-input" name="message-text" placeholder="Введите сообщение..." type="text">
+                </form>
+                <button class="send"><span class="material-symbols-outlined icon">send</span></button>
+                
+            </div>
+            
+
+        </div>
+        
+
+
+        <script>
+            const form = document.querySelector('form');
+            const input = document.querySelector('.form-input');
+            const sendButton = document.querySelector('.send');
+            const containerMe = document.querySelector('.container-me')
+            const containerChat = document.querySelector('.container-chat');
+            window.addEventListener('load', () => {loadMessages();});
+            sendButton.addEventListener('click', handleSubmit);
+            form.addEventListener('submit', this.handleSubmit.bind(this));
+            // localStorage.clear();
+            //form.addEventListener('keypress', this.handleKeyPress.bind(this));
+            function loadMessages(){
+                const messages = [];
+                for (let i = 0; i < localStorage.length; i++){
+                    const key = localStorage.key(i);
+                    const keyDate = new Date(key);
+                    const messageHTML = localStorage.getItem(key);
+                    if (!key.includes('loglevel') && messageHTML) {
+                         messages.push({ key, messageHTML });
+                    }
+                }
+                messages.sort((a, b) => new Date(a.key) - new Date(b.key));
+                for (let i = 0; i < messages.length; i++) {
+                    containerMe.innerHTML += messages[i].messageHTML;
+                    
+                }
+                containerChat.scrollTop = containerChat.scrollHeight;
+                
+            }
+            
+
+            function handleSubmit (event) {
+                event.preventDefault();
+                const messageText = input.value; 
+                if (messageText === '') return;
+                const containerMessage = document.createElement('div');
+                containerMessage.classList.add('container-message');
+                containerMe.appendChild(containerMessage);
+                const nameElement = document.createElement('div');
+                nameElement.classList.add('name');
+                nameElement.innerText = 'Мария Иванова';
+                containerMessage.appendChild(nameElement);
+                const messageElement = document.createElement('div');
+                messageElement.classList.add('message');
+                const timeElement = document.createElement('div');
+                timeElement.classList.add('time');
+                const time = new Date();
+                timeElement.innerText = time.getHours().toString().padStart(2, '0') + ':' + time.getMinutes().toString().padStart(2, '0') + ' ' + time.getDate().toString().padStart(2, '0') + '.'+ time.getMonth().toString().padStart(2, '0') + '.' + time.getFullYear().toString().padStart(2, '0');
+                messageElement.innerText = messageText;
+                input.value = ''; 
+                containerMessage.appendChild(messageElement);
+                containerMessage.appendChild(timeElement);
+                const key = time;
+                const messageData = [nameElement.value, timeElement.value, messageElement.value];
+                const keyTime = time.toISOString();
+                localStorage.setItem(keyTime, containerMessage.outerHTML);
+                containerChat.scrollTop = containerChat.scrollHeight;
+                
+            }
+
+            // function handleKeyPress (event) {
+            //     if (event.keyCode === 13) {
+            //         form.dispatchEvent(new Event('submit'));
+            //     }
+            // }
+        </script>
+    </body>
+</html>
